requirements.txt

streamlit==1.49.1
strands-agents==1.6.0
feedparser==6.0.11
